Renamed DefaultThemeWrapper to be ThemeWrapper. ADded ttf files to be mocked when executing tests.
