+++ src/components/themeWrapper/themeWrapper.tsx
@@ -7,7 +7,7 @@ import { ThemeProvider } from 'emotion-theming';
 import createExtraScopePlugin from 'stylis-plugin-extra-scope';
 
 import defaultTheme from '../../shared/themes/theme';
-import defaultThemeWrapperProps from './defaultThemeWrapper.types';
+import themeWrapperProps from './themeWrapper.types';
 import GlobalStyles from '../Layout/GlobalStyles/GlobalStyles';
 
 const STRONG_ID = 'trova-components-scope';
@@ -16,13 +16,13 @@ const cache = createCache({
 });
 
 /**
- * Renders a <DefaultThemeWrapper> component to wrap other elements of the library and provide
+ * Renders a <ThemeWrapper> component to wrap other elements of the library and provide
  * them the default library theme
  * @param  props
  * @param  props.children - The elements to render inside the theme provider
  */
 
-const DefaultThemeWrapper: React.FC<defaultThemeWrapperProps> = ({
+const ThemeWrapper: React.FC<themeWrapperProps> = ({
     children,
 }) => (
     <CacheProvider value={cache}>
@@ -33,4 +33,4 @@ const DefaultThemeWrapper: React.FC<defaultThemeWrapperProps> = ({
     </CacheProvider>
 );
 
-export default DefaultThemeWrapper;
+export default ThemeWrapper;
