+++ src/index.tsx
@@ -12,7 +12,7 @@ import Toggle from './components/toggle';
 import Checkbox from './components/checkbox';
 import { H1, H2, H3, H4, P, ParagraphSize } from './components/typography';
 import Dropdown, { MultipleDropdown } from './components/dropdown';
-import DefaultThemeWrapper from './components/themeWrapper';
+import ThemeWrapper from './components/themeWrapper';
 
 export {
     Sidebar,
@@ -39,5 +39,5 @@ export {
     ParagraphSize,
     Dropdown,
     MultipleDropdown,
-    DefaultThemeWrapper,
+    ThemeWrapper,
 };
